Adds check for parachain registration files

+++ scripts/register-alphanet-parachain.sh
@@ -12,6 +12,16 @@ if [ -z "$SUDO_SEED" ]; then
     exit 1
 fi
 
+if [ ! -f "$PARACHAIN_WASM" ]; then
+    echo "Missing $PARACHAIN_WASM. Please run scripts/generate-parachain-specs.sh"
+    exit 1
+fi
+
+if [ ! -f "$PARACHAIN_GENESIS" ]; then
+    echo "Missing $PARACHAIN_GENESIS. Please run scripts/generate-parachain-specs.sh"
+    exit 1
+fi
+
 docker run --rm --network=host -v $(pwd)/$PARACHAIN_WASM:/wasm jacogr/polkadot-js-tools api \
     --ws "ws://localhost:$((RELAY_PORT + 2))" \
     --sudo \
